+++ examples/12-materialized/go-purego/materialized/decoder_bounds_test.go
@@ -12,7 +12,7 @@ func TestMaterializedRejectsRootCountBeforeAllocation(t *testing.T) {
 	contracts.RootCounts(t, zigoMaterializedMagicVersion, 1, func(buffer []byte) { zigoDecodeProbeSliceBuffer(buffer) })
 }
 func TestMaterializedRejectsFieldCountBeforeAllocation(t *testing.T) {
-	contracts.FieldCounts(t, zigoMaterializedMagicVersion, 1, 160, []int{72, 88, 152}, func(buffer []byte) { zigoDecodeProbeBuffer(buffer) })
+	contracts.FieldCounts(t, zigoMaterializedMagicVersion, 1, 192, []int{72, 88, 152}, func(buffer []byte) { zigoDecodeProbeBuffer(buffer) })
 }
 func FuzzMaterializedArrayRange(f *testing.F) {
 	contracts.FuzzArrayRange(f, zigoMaterializedArray)